MeSeeGe_Readme
This is a iOS app for typing and checking message when they are on the way home or somewhere. 
This software is developed by Yang Wang(Tucker) and designed by Fei Dong(Ethan), Xinyun(Sherry) Lv is the QE.
Hope you guys like it. It is still under built, wait somtime to get the full edision.